requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
altgraph==0.17.3
async-timeout==4.0.2
attrs==22.2.0
autopep8==2.0.2
cachetools==5.3.0
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
comtypes==1.1.14
frozenlist==1.3.3
google-api-core==2.11.0
google-auth==2.17.3
google-auth-httplib2==0.1.0
google-auth-oauthlib==1.0.0
google-cloud-core==2.3.2
google-cloud-texttospeech==2.14.1
googleapis-common-protos==1.59.0
grpcio==1.53.0
grpcio-status==1.53.0
gTTS==2.3.1
httplib2==0.22.0
idna==3.4
multidict==6.0.4
oauthlib==3.2.2
openai==0.27.4
pefile==2023.2.7
playsound==1.3.0
proto-plus==1.22.2
protobuf==4.22.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
PyAudio==0.2.13
pycodestyle==2.10.0
pyinstaller==5.10.0
pyinstaller-hooks-contrib==2023.2
pyparsing==3.0.9
pypiwin32==223
PyQt5==5.15.9
PyQt5-Qt5==5.15.2
PyQt5-sip==12.12.0
python-dotenv==1.0.0
pyttsx3==2.90
pywin32==306
pywin32-ctypes==0.2.0
requests==2.28.2
requests-oauthlib==1.3.1
rsa==4.9
six==1.16.0
SpeechRecognition==3.10.0
tomli==2.0.1
tqdm==4.65.0
urllib3==1.26.15
yarl==1.8.2
